Block main thread from writing.  Update int to size_t for block_index etc.

+++ src/free_list.h
@@ -33,25 +33,25 @@ namespace mabain {
 
 typedef struct _BufferCache
 {
-    int buf_index;
+    size_t buf_index;
     size_t buf_offset;
 } BufferCache;
 
 class FreeList
 {
 public:
-    FreeList(const std::string &file_path, int buff_alignment, int max_n_buff,
-             int max_buff_per_list = MAX_BUFFER_PER_LIST);
+    FreeList(const std::string &file_path, size_t buff_alignment, size_t max_n_buff,
+             size_t max_buff_per_list = MAX_BUFFER_PER_LIST);
     ~FreeList();
 
     // Free a buffer by adding it to the free list
-    int AddBuffer(size_t offset, int size);
+    int AddBuffer(size_t offset, size_t size);
     // Reserve a buffer by removing it from the free list
-    int RemoveBuffer(size_t &offset, int size);
+    int RemoveBuffer(size_t &offset, size_t size);
     // Release alignment buffer
     void ReleaseAlignmentBuffer(size_t old_offset, size_t alignment_offset);
 
-    bool GetBufferByIndex(int buf_index, size_t &offset);
+    bool GetBufferByIndex(size_t buf_index, size_t &offset);
 
     void Empty();
 
@@ -64,26 +64,26 @@ public:
     // Get total freed buffer size in the list
     size_t GetTotSize() const;
 
-    inline int      AddBufferByIndex(int buf_index, size_t offset);
-    inline size_t   RemoveBufferByIndex(int buf_index);
-    inline int      GetAlignmentSize(int size) const;
-    inline int      GetBufferIndex(int size) const;
-    inline uint64_t GetBufferCountByIndex(int buf_index) const;
-    inline int      GetBufferSizeByIndex(int buf_index) const;
-    inline int      ReleaseBuffer(size_t offset, int size);
+    inline int      AddBufferByIndex(size_t buf_index, size_t offset);
+    inline size_t   RemoveBufferByIndex(size_t buf_index);
+    inline size_t   GetAlignmentSize(size_t size) const;
+    inline size_t   GetBufferIndex(size_t size) const;
+    inline uint64_t GetBufferCountByIndex(size_t buf_index) const;
+    inline size_t   GetBufferSizeByIndex(size_t buf_index) const;
+    inline int      ReleaseBuffer(size_t offset, size_t size);
 
 private:
-    int ReuseBuffer(int buf_index, size_t offset);
+    int ReuseBuffer(size_t buf_index, size_t offset);
 
     // file path where the list will be serialized and stored
     std::string list_path;
     // buffer/memory alignment
-    int alignment;
+    size_t alignment;
     // maximum number of buffers
-    int max_num_buffer;
+    size_t max_num_buffer;
     // maximum buffer per list
     // This restriction is to limit memory usage.
-    int max_buffer_per_list;
+    size_t max_buffer_per_list;
     // number of separated buffers
     MBlsq **buffer_free_list;
     // total count of freed buffers
@@ -92,18 +92,18 @@ private:
     size_t tot_size;
 };
 
-inline int FreeList::GetAlignmentSize(int size) const
+inline size_t FreeList::GetAlignmentSize(size_t size) const
 {
 #ifdef __DEBUG__
     assert(size > 0 && size < max_num_buffer*alignment);
 #endif
-    int alignment_mod = size % alignment;
+    size_t alignment_mod = size % alignment;
     if(alignment_mod == 0)
         return size;
     return (size + alignment - alignment_mod);
 }
 
-inline int FreeList::GetBufferIndex(int size) const
+inline size_t FreeList::GetBufferIndex(size_t size) const
 {
 #ifdef __DEBUG__
     assert(size > 0 && size < max_num_buffer*alignment);
@@ -111,7 +111,7 @@ inline int FreeList::GetBufferIndex(int size) const
     return ((size - 1) / alignment);
 }
 
-inline uint64_t FreeList::GetBufferCountByIndex(int buf_index) const
+inline uint64_t FreeList::GetBufferCountByIndex(size_t buf_index) const
 {
 #ifdef __DEBUG__
     assert(buf_index < max_num_buffer);
@@ -119,7 +119,7 @@ inline uint64_t FreeList::GetBufferCountByIndex(int buf_index) const
     return buffer_free_list[buf_index]->Count();
 }
 
-inline int FreeList::GetBufferSizeByIndex(int buf_index) const
+inline size_t FreeList::GetBufferSizeByIndex(size_t buf_index) const
 {
 #ifdef __DEBUG__
     assert(buf_index < max_num_buffer);
@@ -127,7 +127,7 @@ inline int FreeList::GetBufferSizeByIndex(int buf_index) const
     return (buf_index + 1) * alignment;
 }
 
-inline int FreeList::AddBufferByIndex(int buf_index, size_t offset)
+inline int FreeList::AddBufferByIndex(size_t buf_index, size_t offset)
 {
 #ifdef __DEBUG__
     assert(buf_index < max_num_buffer);
@@ -144,7 +144,7 @@ inline int FreeList::AddBufferByIndex(int buf_index, size_t offset)
     return buffer_free_list[buf_index]->AddIntToTail(offset);
 }
 
-inline size_t FreeList::RemoveBufferByIndex(int buf_index)
+inline size_t FreeList::RemoveBufferByIndex(size_t buf_index)
 {
 #ifdef __DEBUG__
     assert(buf_index < max_num_buffer);
@@ -154,12 +154,12 @@ inline size_t FreeList::RemoveBufferByIndex(int buf_index)
     return buffer_free_list[buf_index]->RemoveIntFromHead();
 }
 
-inline int FreeList::ReleaseBuffer(size_t offset, int size)
+inline int FreeList::ReleaseBuffer(size_t offset, size_t size)
 {
 #ifdef __DEBUG__
     assert(size > 0 && size < max_num_buffer*alignment);
 #endif
-    int buf_index = GetBufferIndex(size);
+    size_t buf_index = GetBufferIndex(size);
     return AddBufferByIndex(buf_index, offset);
 }
